Models now implement their own interfaces.
Models are (by intent) no longer functions.
Problem here is, that it may become necessary,
to deal with situations where the systems of
equations cannot be solved (insufficient rank,
matrix singular etc.). There a mechanism is
needed to handle this and to communicate this
to the surrounding environment without throwing
exceptions. This situation is not exceptional
and can happen.
Also, these edge cases are different for each
model.