Separate model for AffineTransform calc. added to deal with 1D distributed data.
Implementation works but is not yet finished
as I like it. It is also the case of all data
points on a single coordinate (singularity)
missing. Here only translation-x and y will
be passed into the AffineTransform result.

This singularity special case also must be
added to the rigid transform model as well.